+++ packages/forge-rhf/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nexcraft/forge-rhf
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @nexcraft/forge@0.9.0
+
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @nexcraft/forge@0.8.0
+
 ## 0.3.0
 
 ### Minor Changes